Refactor layout code
